remove wellOrder when collapsed; add adv settings shaded region; update padding/margins
